requirements.txt
hydra-core==1.0.3
torch==1.6.0
torchvision==0.7.0
wandb==0.10.5